+++ R/filter_by_shape_id.R
@@ -52,18 +52,18 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
   # 'shapes' (shape_id)
 
-  if (gtfsio::check_fields_exist(gtfs, "shapes", "shape_id")) {
+  if (gtfsio::check_field_exists(gtfs, "shapes", "shape_id")) {
 
-    gtfsio::assert_fields_types(gtfs, "shapes", "shape_id", "character")
+    gtfsio::assert_field_class(gtfs, "shapes", "shape_id", "character")
     gtfs$shapes <- gtfs$shapes[shape_id %ffilter% get("shape_id", envir = env)]
 
   }
 
   # 'trips' (shape_id)
 
-  if (gtfsio::check_fields_exist(gtfs, "trips", "shape_id")) {
+  if (gtfsio::check_field_exists(gtfs, "trips", "shape_id")) {
 
-    gtfsio::assert_fields_types(gtfs, "trips", "shape_id", "character")
+    gtfsio::assert_field_class(gtfs, "trips", "shape_id", "character")
     gtfs$trips <- gtfs$trips[shape_id %ffilter% get("shape_id", envir = env)]
 
     # 'trips' allows us to filter by 'route_id', 'service_id' and 'trip_id'
@@ -74,16 +74,16 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
     # 'calendar' and 'calendar_dates' (service_id)
 
-    if (gtfsio::check_fields_exist(gtfs, "calendar", "service_id")) {
+    if (gtfsio::check_field_exists(gtfs, "calendar", "service_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "calendar", "service_id", "character")
+      gtfsio::assert_field_class(gtfs, "calendar", "service_id", "character")
       gtfs$calendar <- gtfs$calendar[service_id %chin% relevant_services]
 
     }
 
-    if (gtfsio::check_fields_exist(gtfs, "calendar_dates", "service_id")) {
+    if (gtfsio::check_field_exists(gtfs, "calendar_dates", "service_id")) {
 
-      gtfsio::assert_fields_types(
+      gtfsio::assert_field_class(
         gtfs,
         "calendar_dates",
         "service_id",
@@ -103,9 +103,9 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
     relevant_agencies <- vector("character", length = 0L)
 
-    if (gtfsio::check_fields_exist(gtfs, "routes", "route_id")) {
+    if (gtfsio::check_field_exists(gtfs, "routes", "route_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "routes", "route_id", "character")
+      gtfsio::assert_field_class(gtfs, "routes", "route_id", "character")
       gtfs$routes <- gtfs$routes[route_id %chin% relevant_routes]
 
       # 'routes' allows us to filter by 'agency_id'.
@@ -121,9 +121,9 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
     }
 
-    if (gtfsio::check_fields_exist(gtfs, "fare_rules", "route_id")) {
+    if (gtfsio::check_field_exists(gtfs, "fare_rules", "route_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "fare_rules", "route_id", "character")
+      gtfsio::assert_field_class(gtfs, "fare_rules", "route_id", "character")
       gtfs$fare_rules <- gtfs$fare_rules[route_id %chin% relevant_routes]
 
       # 'fare_rules' allows us to filter by 'fare_id'
@@ -132,9 +132,9 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
       # 'fare_attributes' (fare_id)
 
-      if (gtfsio::check_fields_exist(gtfs, "fare_attributes", "fare_id")) {
+      if (gtfsio::check_field_exists(gtfs, "fare_attributes", "fare_id")) {
 
-        gtfsio::assert_fields_types(
+        gtfsio::assert_field_class(
           gtfs,
           "fare_attributes",
           "fare_id",
@@ -161,30 +161,30 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
     # 'agency' (agency_id, that comes both from routes and fare_attributes)
 
-    if (gtfsio::check_fields_exist(gtfs, "agency", "agency_id") &&
+    if (gtfsio::check_field_exists(gtfs, "agency", "agency_id") &&
         length(relevant_agencies) > 0) {
 
       # keeping only unique agency_ids from relevant_agencies, since they may
       # come from two different sources
       relevant_agencies <- unique(relevant_agencies)
 
-      gtfsio::assert_fields_types(gtfs, "agency", "agency_id", "character")
+      gtfsio::assert_field_class(gtfs, "agency", "agency_id", "character")
       gtfs$agency <- gtfs$agency[agency_id %chin% relevant_agencies]
 
     }
 
     # 'frequencies' and 'stop_times' (trip_id)
 
-    if (gtfsio::check_fields_exist(gtfs, "frequencies", "trip_id")) {
+    if (gtfsio::check_field_exists(gtfs, "frequencies", "trip_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "frequencies", "trip_id", "character")
+      gtfsio::assert_field_class(gtfs, "frequencies", "trip_id", "character")
       gtfs$frequencies <- gtfs$frequencies[trip_id %chin% relevant_trips]
 
     }
 
-    if (gtfsio::check_fields_exist(gtfs, "stop_times", "trip_id")) {
+    if (gtfsio::check_field_exists(gtfs, "stop_times", "trip_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "stop_times", "trip_id", "character")
+      gtfsio::assert_field_class(gtfs, "stop_times", "trip_id", "character")
       gtfs$stop_times <- gtfs$stop_times[trip_id %chin% relevant_trips]
 
       # 'stop_times' allows us to filter by 'stop_id'. it's important to keep,
@@ -193,7 +193,7 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
       relevant_stops <- unique(gtfs$stop_times$stop_id)
 
-      if (gtfsio::check_fields_exist(gtfs, "stops", "parent_station")) {
+      if (gtfsio::check_field_exists(gtfs, "stops", "parent_station")) {
 
         # this may result in a warning if stop_times has stop_ids not listed in
         # stops
@@ -206,9 +206,9 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
 
       # 'stops' (stop_id)
 
-      if (gtfsio::check_fields_exist(gtfs, "stops", "stop_id")) {
+      if (gtfsio::check_field_exists(gtfs, "stops", "stop_id")) {
 
-        gtfsio::assert_fields_types(gtfs, "stops", "stop_id", "character")
+        gtfsio::assert_field_class(gtfs, "stops", "stop_id", "character")
         gtfs$stops <- gtfs$stops[stop_id %chin% relevant_stops]
 
       }