+++ R/filter_by_trip_id.R
@@ -52,9 +52,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
   # 'trips' (trip_id)
 
-  if (gtfsio::check_fields_exist(gtfs, "trips", "trip_id")) {
+  if (gtfsio::check_field_exists(gtfs, "trips", "trip_id")) {
 
-    gtfsio::assert_fields_types(gtfs, "trips", "trip_id", "character")
+    gtfsio::assert_field_class(gtfs, "trips", "trip_id", "character")
     gtfs$trips <- gtfs$trips[trip_id %ffilter% get("trip_id", envir = env)]
 
     # 'trips' allows us to filter by 'route_id', 'service_id' and 'shape_id'
@@ -65,25 +65,25 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     # 'shapes' (shape_id)
 
-    if (gtfsio::check_fields_exist(gtfs, "shapes", "shape_id")) {
+    if (gtfsio::check_field_exists(gtfs, "shapes", "shape_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "shapes", "shape_id", "character")
+      gtfsio::assert_field_class(gtfs, "shapes", "shape_id", "character")
       gtfs$shapes <- gtfs$shapes[shape_id %chin% relevant_shapes]
 
     }
 
     # 'calendar' and 'calendar_dates' (service_id)
 
-    if (gtfsio::check_fields_exist(gtfs, "calendar", "service_id")) {
+    if (gtfsio::check_field_exists(gtfs, "calendar", "service_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "calendar", "service_id", "character")
+      gtfsio::assert_field_class(gtfs, "calendar", "service_id", "character")
       gtfs$calendar <- gtfs$calendar[service_id %chin% relevant_services]
 
     }
 
-    if (gtfsio::check_fields_exist(gtfs, "calendar_dates", "service_id")) {
+    if (gtfsio::check_field_exists(gtfs, "calendar_dates", "service_id")) {
 
-      gtfsio::assert_fields_types(
+      gtfsio::assert_field_class(
         gtfs,
         "calendar_dates",
         "service_id",
@@ -104,9 +104,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     relevant_agencies <- vector("character", length = 0L)
 
-    if (gtfsio::check_fields_exist(gtfs, "routes", "route_id")) {
+    if (gtfsio::check_field_exists(gtfs, "routes", "route_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "routes", "route_id", "character")
+      gtfsio::assert_field_class(gtfs, "routes", "route_id", "character")
       gtfs$routes <- gtfs$routes[route_id %chin% relevant_routes]
 
       # 'routes' allows us to filter by 'agency_id'. but 'agency_id' is
@@ -124,9 +124,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     }
 
-    if (gtfsio::check_fields_exist(gtfs, "fare_rules", "route_id")) {
+    if (gtfsio::check_field_exists(gtfs, "fare_rules", "route_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "fare_rules", "route_id", "character")
+      gtfsio::assert_field_class(gtfs, "fare_rules", "route_id", "character")
       gtfs$fare_rules <- gtfs$fare_rules[route_id %chin% relevant_routes]
 
       # 'fare_rules' allows us to filter by 'fare_id'
@@ -135,9 +135,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
       # 'fare_attributes' (fare_id)
 
-      if (gtfsio::check_fields_exist(gtfs, "fare_attributes", "fare_id")) {
+      if (gtfsio::check_field_exists(gtfs, "fare_attributes", "fare_id")) {
 
-        gtfsio::assert_fields_types(
+        gtfsio::assert_field_class(
           gtfs,
           "fare_attributes",
           "fare_id",
@@ -164,14 +164,14 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     # 'agency' (agency_id, that comes both from routes and fare_attributes)
 
-    if (gtfsio::check_fields_exist(gtfs, "agency", "agency_id") &
+    if (gtfsio::check_field_exists(gtfs, "agency", "agency_id") &
         length(relevant_agencies) > 0) {
 
       # keeping only unique agency_ids from relevant_agencies, since it may come
       # from two differente sources
       relevant_agencies <- unique(relevant_agencies)
 
-      gtfsio::assert_fields_types(gtfs, "agency", "agency_id", "character")
+      gtfsio::assert_field_class(gtfs, "agency", "agency_id", "character")
       gtfs$agency <- gtfs$agency[agency_id %chin% relevant_agencies]
 
     }
@@ -180,9 +180,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
   # 'stop_times' (trip_id)
 
-  if (gtfsio::check_fields_exist(gtfs, "stop_times", "trip_id")) {
+  if (gtfsio::check_field_exists(gtfs, "stop_times", "trip_id")) {
 
-    gtfsio::assert_fields_types(gtfs, "stop_times", "trip_id", "character")
+    gtfsio::assert_field_class(gtfs, "stop_times", "trip_id", "character")
     gtfs$stop_times <- gtfs$stop_times[
       trip_id %ffilter% get("trip_id", envir = env)
     ]
@@ -193,7 +193,7 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     relevant_stops <- unique(gtfs$stop_times$stop_id)
 
-    if (gtfsio::check_fields_exist(gtfs, "stops", "parent_station")) {
+    if (gtfsio::check_field_exists(gtfs, "stops", "parent_station")) {
 
       suppressWarnings(
         stops_with_parents <- get_parent_station(gtfs, relevant_stops)
@@ -206,9 +206,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     from_to_stop_id <- c("from_stop_id", "to_stop_id")
 
-    if (gtfsio::check_fields_exist(gtfs, "transfers", from_to_stop_id)) {
+    if (gtfsio::check_field_exists(gtfs, "transfers", from_to_stop_id)) {
 
-      gtfsio::assert_fields_types(
+      gtfsio::assert_field_class(
         gtfs,
         "transfers",
         from_to_stop_id,
@@ -220,9 +220,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     }
 
-    if (gtfsio::check_fields_exist(gtfs, "pathways", from_to_stop_id)) {
+    if (gtfsio::check_field_exists(gtfs, "pathways", from_to_stop_id)) {
 
-      gtfsio::assert_fields_types(
+      gtfsio::assert_field_class(
         gtfs,
         "pathways",
         from_to_stop_id,
@@ -234,9 +234,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
     }
 
-    if (gtfsio::check_fields_exist(gtfs, "stops", "stop_id")) {
+    if (gtfsio::check_field_exists(gtfs, "stops", "stop_id")) {
 
-      gtfsio::assert_fields_types(gtfs, "stops", "stop_id", "character")
+      gtfsio::assert_field_class(gtfs, "stops", "stop_id", "character")
       gtfs$stops <- gtfs$stops[stop_id %chin% relevant_stops]
 
       # 'stops' allows us to filter by 'level_id'
@@ -245,9 +245,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
       # 'levels' (level_id)
 
-      if (gtfsio::check_fields_exist(gtfs, "levels", "level_id")) {
+      if (gtfsio::check_field_exists(gtfs, "levels", "level_id")) {
 
-        gtfsio::assert_fields_types(gtfs, "levels", "level_id", "character")
+        gtfsio::assert_field_class(gtfs, "levels", "level_id", "character")
         gtfs$levels <- gtfs$levels[level_id %chin% relevant_levels]
 
       }
@@ -258,9 +258,9 @@ filter_by_trip_id <- function(gtfs, trip_id, keep = TRUE) {
 
   # 'frequencies' (trip_id)
 
-  if (gtfsio::check_fields_exist(gtfs, "frequencies", "trip_id")) {
+  if (gtfsio::check_field_exists(gtfs, "frequencies", "trip_id")) {
 
-    gtfsio::assert_fields_types(gtfs, "frequencies", "trip_id", "character")
+    gtfsio::assert_field_class(gtfs, "frequencies", "trip_id", "character")
     gtfs$frequencies <- gtfs$frequencies[
       trip_id %ffilter% get("trip_id", envir = env)
     ]