+++ R/get_trip_geometry.R
@@ -74,13 +74,13 @@ get_trip_geometry <- function(gtfs,
 
   if ("shapes" %in% file) {
 
-    gtfsio::assert_fields_types(
+    gtfsio::assert_field_class(
       gtfs,
       "trips",
       c("trip_id", "shape_id"),
       rep("character", 2)
     )
-    gtfsio::check_fields_types(
+    gtfsio::assert_field_class(
       gtfs,
       "shapes",
       c("shape_id", "shape_pt_lat", "shape_pt_lon", "shape_pt_sequence"),
@@ -91,14 +91,14 @@ get_trip_geometry <- function(gtfs,
 
   if ("stop_times" %in% file) {
 
-    gtfsio::assert_fields_types(gtfs, "trips", "trip_id", "character")
-    gtfsio::assert_fields_types(
+    gtfsio::assert_field_class(gtfs, "trips", "trip_id", "character")
+    gtfsio::assert_field_class(
       gtfs,
       "stop_times",
       c("trip_id", "stop_id", "stop_sequence"),
       c("character", "character", "integer")
     )
-    gtfsio::assert_fields_types(
+    gtfsio::assert_field_class(
       gtfs,
       "stops",
       c("stop_id", "stop_lat", "stop_lon"),