+++ R/get_trip_segment_duration.R
@@ -47,7 +47,7 @@ get_trip_segment_duration <- function(gtfs, trip_id = NULL, unit = "min") {
 
   # check if required fields and files exist
 
-  gtfsio::assert_fields_types(
+  gtfsio::assert_field_class(
     gtfs,
     "stop_times",
     c("trip_id", "arrival_time", "departure_time", "stop_sequence"),